backend/requirements.txt
# BlogCMS Backend Dependencies - Minimal Setup

# ============================================================================
# CORE FRAMEWORK
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# ============================================================================
# DATABASE
# ============================================================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.0
slowapi==0.1.9  # Rate limiting
Pillow==12.0.0  # Image processing for media uploads

# ============================================================================
# DEVELOPMENT (Optional)
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
